pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "httpx>=0.28.1",
    "pdfplumber==0.10.3",
    "pillow==10.1.0",
    "pydantic==2.5.0",
    "pytesseract==0.3.10",
    "pytest>=8.4.2",
    "python-multipart==0.0.6",
    "uvicorn==0.24.0",
]
